requirements.txt
pywebview[qt]==6.2.1
